доработка структуры
