+++ project/src/components/private-route/index.tsx
