+++ project/src/pages/main-empty/index.tsx
@@ -1,4 +1,4 @@
-import Header from '../../components/header/header';
+import Header from '../../components/header';
 
 const MainEmpty = () => (
   <div className="page page--gray page--main">
